+++ piecrust/CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add crate-specific README. [#174]
 
+### Changed
+
+- Change `owner` parameter type in `ModuleData::builder` to be `[u8; N]` [#201] 
+
 ## [0.3.0] - 2023-04-26
 
 ### Changed
@@ -50,6 +54,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - First `piecrust` release
 
 <!-- ISSUES -->
+[#201]: https://github.com/dusk-network/piecrust/issues/201
 [#181]: https://github.com/dusk-network/piecrust/issues/181
 [#178]: https://github.com/dusk-network/piecrust/issues/178
 [#174]: https://github.com/dusk-network/piecrust/issues/174